ETL/requirements.txt
python-dotenv==0.19.2
tqdm==4.66.1
psycopg2-binary==2.9.6